refactor(broker-core): extract first broker kernel slice
