+++ broker-core/package.json
@@ -1,0 +1,46 @@
+{
+  "name": "@gugu910/pi-broker-core",
+  "version": "0.0.0",
+  "type": "module",
+  "description": "Transport-neutral broker kernel primitives for pi transports",
+  "author": "Will Porcellini <5994936+gugu91@users.noreply.github.com>",
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/gugu91/extensions.git",
+    "directory": "broker-core"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "main": "./dist/index.js",
+  "exports": {
+    ".": "./dist/index.js",
+    "./agent-messaging": "./dist/agent-messaging.js",
+    "./auth": "./dist/auth.js",
+    "./leader": "./dist/leader.js",
+    "./maintenance": "./dist/maintenance.js",
+    "./message-send": "./dist/message-send.js",
+    "./paths": "./dist/paths.js",
+    "./raw-tcp-loopback": "./dist/raw-tcp-loopback.js",
+    "./router": "./dist/router.js",
+    "./schema": "./dist/schema.js",
+    "./types": "./dist/types.js",
+    "./package.json": "./package.json"
+  },
+  "files": [
+    "README.md",
+    "LICENSE",
+    "dist/"
+  ],
+  "pi": {},
+  "scripts": {
+    "build": "node ../scripts/build-package.mjs",
+    "prepack": "pnpm run build",
+    "lint": "eslint . --ext .ts",
+    "typecheck": "tsc --noEmit"
+  },
+  "dependencies": {
+    "@gugu910/pi-transport-core": "workspace:*"
+  }
+}
