+++ scripts/build-package.mjs
@@ -16,6 +16,13 @@ const packageConfigs = {
     vendorDirs: [],
     importRewrites: [],
   },
+  "broker-core": {
+    excludeDirs: new Set(["dist", "node_modules", ".turbo"]),
+    excludeFiles: new Set(),
+    excludePrefixes: [],
+    vendorDirs: [],
+    importRewrites: [],
+  },
   "imessage-bridge": {
     excludeDirs: new Set(["dist", "node_modules", ".turbo"]),
     excludeFiles: new Set(),
