+++ slack-bridge/broker/router.ts
@@ -1,370 +1,1 @@
-import { buildPinetOwnerToken } from "../helpers.js";
-import type { AgentInfo, BrokerDBInterface, InboundMessage, RoutingDecision } from "./types.js";
-
-// ─── Helpers ─────────────────────────────────────────────
-
-export interface ThreadOwnerHint {
-  agentOwner?: string;
-  agentName?: string;
-}
-
-export type ExplicitThreadDirective =
-  | { kind: "stand_down"; agent: AgentInfo }
-  | { kind: "retarget"; agent: AgentInfo };
-
-interface AgentMentionCandidate {
-  agent: AgentInfo;
-  term: string;
-}
-
-const THREAD_STAND_DOWN_REGEX = /\bstand down\b/i;
-const THREAD_RETARGET_REGEX =
-  /\b(?:take over|pick (?:this|it) up|pick this up|pick it up|handle this|grab this|you take this|reassign(?: this)?(?: to)?|switch(?: this)?(?: to)?|move(?: this)?(?: to)?|route(?: this)?(?: to)?|transfer(?: this)?(?: to)?|pass(?: this)?(?: to)?|hand(?: this)?(?: to)?|give(?: this)?(?: to)?)\b/i;
-
-/**
- * Extract an agent name mention from message text.
- * Matches patterns like "hey AgentName," or "@AgentName" or just "AgentName"
- * at word boundaries (case-insensitive).
- *
- * When multiple agents match, the longest name wins so that "CodeBot" is
- * preferred over "Code" and similar-prefix collisions are avoided.
- */
-export function findAgentMention(text: string, agents: AgentInfo[]): AgentInfo | null {
-  return findBestAgentMention(text, buildAgentMentionCandidates(agents, false));
-}
-
-export function extractPiAgentThreadOwnerHint(
-  replies: ReadonlyArray<Record<string, unknown>>,
-): ThreadOwnerHint | null {
-  for (let index = replies.length - 1; index >= 0; index -= 1) {
-    const message = replies[index];
-    if (!message.bot_id) continue;
-    const metadata = message.metadata as
-      | {
-          event_type?: string;
-          event_payload?: { agent?: string; agent_owner?: string };
-        }
-      | undefined;
-    if (metadata?.event_type !== "pi_agent_msg") continue;
-
-    const agentOwner =
-      typeof metadata.event_payload?.agent_owner === "string" &&
-      metadata.event_payload.agent_owner.trim().length > 0
-        ? metadata.event_payload.agent_owner.trim()
-        : undefined;
-    const agentName =
-      typeof metadata.event_payload?.agent === "string" &&
-      metadata.event_payload.agent.trim().length > 0
-        ? metadata.event_payload.agent.trim()
-        : undefined;
-
-    if (agentOwner || agentName) {
-      return {
-        ...(agentOwner ? { agentOwner } : {}),
-        ...(agentName ? { agentName } : {}),
-      };
-    }
-  }
-  return null;
-}
-
-export function findExplicitThreadDirective(
-  text: string,
-  agents: AgentInfo[],
-): ExplicitThreadDirective | null {
-  if (!THREAD_STAND_DOWN_REGEX.test(text) && !THREAD_RETARGET_REGEX.test(text)) {
-    return null;
-  }
-
-  const mention = findBestAgentMention(text, buildAgentMentionCandidates(agents, true));
-  if (!mention) return null;
-
-  if (THREAD_STAND_DOWN_REGEX.test(text)) {
-    return { kind: "stand_down", agent: mention };
-  }
-
-  if (THREAD_RETARGET_REGEX.test(text)) {
-    return { kind: "retarget", agent: mention };
-  }
-
-  return null;
-}
-
-function buildAgentMentionCandidates(
-  agents: AgentInfo[],
-  includeUniqueTailAlias: boolean,
-): AgentMentionCandidate[] {
-  const candidates: AgentMentionCandidate[] = [];
-  const tailCounts = new Map<string, number>();
-
-  for (const agent of agents) {
-    const tail = getAgentTailAlias(agent.name);
-    if (tail) {
-      tailCounts.set(tail, (tailCounts.get(tail) ?? 0) + 1);
-    }
-  }
-
-  for (const agent of agents) {
-    if (agent.name.trim()) {
-      candidates.push({ agent, term: agent.name.trim() });
-    }
-
-    if (!includeUniqueTailAlias) continue;
-    const tail = getAgentTailAlias(agent.name);
-    if (!tail || tailCounts.get(tail) !== 1) continue;
-    candidates.push({ agent, term: tail });
-  }
-
-  return candidates;
-}
-
-function getAgentTailAlias(name: string): string | null {
-  const tokens = name
-    .trim()
-    .split(/\s+/)
-    .map((token) => token.trim())
-    .filter(Boolean);
-  const tail = tokens.at(-1)?.toLowerCase() ?? "";
-  return tail.length >= 4 ? tail : null;
-}
-
-function findBestAgentMention(text: string, candidates: AgentMentionCandidate[]): AgentInfo | null {
-  const lower = text.toLowerCase();
-  let bestMatch: AgentInfo | null = null;
-  let bestLength = 0;
-
-  for (const candidate of candidates) {
-    const term = candidate.term.toLowerCase();
-    if (!term) continue;
-    const pattern = new RegExp(`\\b${escapeRegExp(term)}\\b`, "i");
-    if (pattern.test(lower) && term.length > bestLength) {
-      bestMatch = candidate.agent;
-      bestLength = term.length;
-    }
-  }
-
-  return bestMatch;
-}
-
-function resolveAgentFromThreadOwnerHint(
-  metadata: Record<string, unknown> | undefined,
-  agents: AgentInfo[],
-): AgentInfo | null {
-  const hintedOwner =
-    typeof metadata?.threadOwnerAgentOwner === "string" && metadata.threadOwnerAgentOwner
-      ? metadata.threadOwnerAgentOwner
-      : null;
-  if (hintedOwner) {
-    const ownerMatch = agents.find(
-      (agent) => agent.stableId && buildPinetOwnerToken(agent.stableId) === hintedOwner,
-    );
-    if (ownerMatch) {
-      return ownerMatch;
-    }
-  }
-
-  const hintedName =
-    typeof metadata?.threadOwnerAgentName === "string" && metadata.threadOwnerAgentName
-      ? metadata.threadOwnerAgentName
-      : null;
-  if (!hintedName) {
-    return null;
-  }
-
-  return findBestAgentMention(hintedName, buildAgentMentionCandidates(agents, false));
-}
-
-function resolveRoutableThreadOwner(
-  db: BrokerDBInterface,
-  threadOwnerAgentId: string | null,
-  now = new Date().toISOString(),
-): AgentInfo | null {
-  if (!threadOwnerAgentId) return null;
-
-  const owner = db.getAgentById(threadOwnerAgentId);
-  if (owner && isRoutableOwner(owner, now)) {
-    return owner;
-  }
-
-  if (!owner) {
-    const reconnectedOwner = db.getAgentByStableId(threadOwnerAgentId);
-    if (
-      reconnectedOwner &&
-      reconnectedOwner.id !== threadOwnerAgentId &&
-      isRoutableOwner(reconnectedOwner, now)
-    ) {
-      return reconnectedOwner;
-    }
-  }
-
-  return null;
-}
-
-function escapeRegExp(s: string): string {
-  return s.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
-}
-
-function isRoutableOwner(agent: AgentInfo, now = new Date().toISOString()): boolean {
-  if (!agent.disconnectedAt) return true;
-  return agent.resumableUntil != null && agent.resumableUntil > now;
-}
-
-// ─── MessageRouter ───────────────────────────────────────
-
-export class MessageRouter {
-  private readonly db: BrokerDBInterface;
-
-  constructor(db: BrokerDBInterface) {
-    this.db = db;
-  }
-
-  /**
-   * Route an inbound message to the right agent.
-   *
-   * Priority order:
-   * 1. User allowlist — reject if user not allowed
-   * 2. Explicit thread control — stand-down / reassignment signals in-thread
-   * 3. Known-thread ownership — authoritative owner hint, then broker DB owner
-   * 4. New-thread channel assignment / direct address
-   * 5. Unrouted — no match found
-   */
-  route(msg: InboundMessage): RoutingDecision {
-    // 0. Check user allowlist
-    const allowedUsers = this.db.getAllowedUsers();
-    if (allowedUsers !== null && !allowedUsers.has(msg.userId)) {
-      return { action: "reject", reason: "User not in allowlist" };
-    }
-
-    const agents = this.db.getAgents();
-    let thread = this.db.getThread(msg.threadId);
-    const explicitDirective = findExplicitThreadDirective(msg.text, agents);
-
-    if (explicitDirective) {
-      if (thread) {
-        if (explicitDirective.kind === "retarget") {
-          this.db.updateThread(msg.threadId, {
-            ownerAgent: explicitDirective.agent.id,
-            ownerBinding: "explicit",
-            channel: msg.channel,
-            source: msg.source,
-          });
-        }
-        return { action: "deliver", agentId: explicitDirective.agent.id };
-      }
-    }
-
-    if (thread) {
-      if (thread.ownerBinding === "explicit") {
-        const explicitOwner = resolveRoutableThreadOwner(this.db, thread.ownerAgent);
-        if (explicitOwner) {
-          if (thread.ownerAgent !== explicitOwner.id) {
-            this.db.updateThread(msg.threadId, { ownerAgent: explicitOwner.id });
-          }
-          return { action: "deliver", agentId: explicitOwner.id };
-        }
-
-        if (thread.ownerAgent !== null) {
-          this.db.updateThread(msg.threadId, { ownerAgent: null });
-        }
-
-        // Explicit takeovers stay authoritative for the thread. If that owner is
-        // unavailable later, require another explicit retarget instead of snapping
-        // back to a stale historical Slack owner hint.
-        return { action: "unrouted" };
-      }
-
-      const hintedOwner = resolveAgentFromThreadOwnerHint(msg.metadata, agents);
-      if (hintedOwner && isRoutableOwner(hintedOwner)) {
-        if (thread.ownerAgent !== hintedOwner.id) {
-          this.db.updateThread(msg.threadId, { ownerAgent: hintedOwner.id, channel: msg.channel });
-        }
-        return { action: "deliver", agentId: hintedOwner.id };
-      }
-
-      if (thread.ownerAgent) {
-        const owner = resolveRoutableThreadOwner(this.db, thread.ownerAgent);
-        if (owner) {
-          if (thread.ownerAgent !== owner.id) {
-            this.db.updateThread(msg.threadId, { ownerAgent: owner.id });
-          }
-          return { action: "deliver", agentId: owner.id };
-        }
-
-        // Owner is gone or no longer routable — clear ownership so the thread can
-        // be explicitly rebound, but do not leak a known-thread reply to another
-        // worker through generic fallback routing.
-        this.db.updateThread(msg.threadId, { ownerAgent: null });
-        thread = this.db.getThread(msg.threadId);
-      }
-
-      const mentioned = findAgentMention(msg.text, agents);
-      if (mentioned) {
-        const claimed = this.db.claimThread(msg.threadId, mentioned.id, msg.source, msg.channel);
-        if (claimed) {
-          return { action: "deliver", agentId: mentioned.id };
-        }
-
-        const claimedThread = this.db.getThread(msg.threadId);
-        const claimedOwner = resolveRoutableThreadOwner(this.db, claimedThread?.ownerAgent ?? null);
-        if (claimedOwner) {
-          return { action: "deliver", agentId: claimedOwner.id };
-        }
-      }
-
-      return { action: "unrouted" };
-    }
-
-    // New thread / top-level message: channel assignment can still steer work.
-    const assignment = this.db.getChannelAssignment(msg.channel);
-    if (assignment) {
-      const assigned = agents.find((agent) => agent.id === assignment.agentId);
-      if (assigned) {
-        return { action: "deliver", agentId: assigned.id };
-      }
-    }
-
-    const mentioned = findAgentMention(msg.text, agents);
-    if (mentioned) {
-      const claimed = this.db.claimThread(msg.threadId, mentioned.id, msg.source, msg.channel);
-      if (claimed) {
-        return { action: "deliver", agentId: mentioned.id };
-      }
-
-      const claimedThread = this.db.getThread(msg.threadId);
-      const claimedOwner = resolveRoutableThreadOwner(this.db, claimedThread?.ownerAgent ?? null);
-      if (claimedOwner) {
-        return { action: "deliver", agentId: claimedOwner.id };
-      }
-    }
-
-    return { action: "unrouted" };
-  }
-
-  /**
-   * Claim a thread for an agent (first-responder-wins).
-   * Optionally provide the transport source and channel to store when creating
-   * a new thread. Defaults to Slack for backward compatibility.
-   * Returns true if the claim succeeded, false if another agent already owns it.
-   *
-   * Delegates to the DB layer which performs the claim atomically
-   * (single SQL statement) to avoid TOCTOU races. (#125)
-   */
-  claimThread(threadId: string, agentId: string, channel?: string, source = "slack"): boolean {
-    return this.db.claimThread(threadId, agentId, source, channel ?? "");
-  }
-
-  /**
-   * Get the owner of a thread, or null if unclaimed / nonexistent.
-   */
-  getThreadOwner(threadId: string): string | null {
-    const thread = this.db.getThread(threadId);
-    return thread?.ownerAgent ?? null;
-  }
-
-  /**
-   * List available (connected) agents for routing.
-   */
-  getAvailableAgents(): AgentInfo[] {
-    return this.db.getAgents();
-  }
-}
+export * from "@gugu910/pi-broker-core";
