+++ slack-bridge/package.json
@@ -40,6 +40,7 @@
     "test": "vitest run"
   },
   "dependencies": {
+    "@gugu910/pi-broker-core": "workspace:*",
     "@gugu910/pi-imessage-bridge": "workspace:*",
     "@gugu910/pi-transport-core": "workspace:*",
     "@sinclair/typebox": "^0.34.49"
